add casts to silent compiler warnings with 64bit systems.

+++ lib/socks.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: socks.c,v 1.31 2009-07-09 21:47:24 bagder Exp $
+ * $Id: socks.c,v 1.32 2009-08-29 03:53:09 gknauf Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -514,10 +514,10 @@ CURLcode Curl_SOCKS5(const char *proxy_name,
     socksreq[len++] = 1;    /* username/pw subnegotiation version */
     socksreq[len++] = (char) userlen;
     memcpy(socksreq + len, proxy_name, (int) userlen);
-    len += userlen;
+    len += (int)userlen;
     socksreq[len++] = (char) pwlen;
     memcpy(socksreq + len, proxy_password, (int) pwlen);
-    len += pwlen;
+    len += (int)pwlen;
 
     code = Curl_write_plain(conn, sock, (char *)socksreq, len, &written);
     if((code != CURLE_OK) || (len != written)) {
